requirements.txt
torch
torchvision
numpy<2
pandas
matplotlib
scikit-learn
pillow
seaborn
tqdm
jupyter
pyyaml
onnx
onnxruntime
